remove everything from address_manager_store, except `PeerDataSerialization`
