+++ benchmarks/address_manager_store.py
@@ -17,7 +17,6 @@ from chia.server.address_manager import (
     AddressManager,
     ExtendedPeerInfo,
 )
-from chia.server.address_manager_store import deserialize_bytes, serialize_bytes
 from chia.types.peer_info import TimestampedPeerInfo
 from chia.util.files import write_file_async
 
@@ -103,7 +102,7 @@ async def benchmark_serialize_deserialize(iterations: int = 5) -> None:
             # Benchmark serialize
             start_serialize = time.perf_counter()
 
-            serialised_bytes = serialize_bytes(address_manager)
+            serialised_bytes = address_manager.serialize_bytes()
             await write_file_async(peers_file_path, serialised_bytes, file_mode=0o644)
             end_serialize = time.perf_counter()
             serialize_duration = end_serialize - start_serialize
@@ -114,7 +113,7 @@ async def benchmark_serialize_deserialize(iterations: int = 5) -> None:
             async with aiofiles.open(peers_file_path, "rb") as f:
                 data = io.BytesIO(await f.read())
             start_deserialize = time.perf_counter()
-            _ = deserialize_bytes(data)
+            _ = AddressManager.deserialize_bytes(data)
             end_deserialize = time.perf_counter()
             deserialize_duration = end_deserialize - start_deserialize
             total_deserialize_time += deserialize_duration
