+++ chia/server/node_discovery.py
@@ -20,7 +20,6 @@ from chia.protocols.full_node_protocol import RequestPeers, RespondPeers
 from chia.protocols.introducer_protocol import RequestPeersIntroducer
 from chia.protocols.protocol_message_types import ProtocolMessageTypes
 from chia.server.address_manager import AddressManager, ExtendedPeerInfo
-from chia.server.address_manager_store import AddressManagerStore, serialize_bytes
 from chia.server.outbound_message import Message, NodeType, make_msg
 from chia.server.server import ChiaServer
 from chia.server.ws_connection import WSChiaConnection
@@ -88,7 +87,7 @@ class FullNodeDiscovery:
             self.default_port = NETWORK_ID_DEFAULT_PORTS[self.selected_network]
 
     async def initialize_address_manager(self) -> None:
-        self.address_manager = await AddressManagerStore.create_address_manager(self.peers_file_path)
+        self.address_manager = await AddressManager.create_address_manager(self.peers_file_path)
         if self.enable_private_networks:
             self.address_manager.make_private_subnets_valid()
         self.server.set_received_message_callback(self.update_peer_timestamp_on_message)
@@ -415,7 +414,7 @@ class FullNodeDiscovery:
             serialize_interval = random.randint(15 * 60, 30 * 60)
             await asyncio.sleep(serialize_interval)
             async with self.address_manager.lock:
-                serialised_bytes = serialize_bytes(self.address_manager)
+                serialised_bytes = self.address_manager.serialize_bytes()
             await write_file_async(self.peers_file_path, serialised_bytes, file_mode=0o644)
 
     async def _periodically_cleanup(self) -> None:
